+++ src/game_state.rs
@@ -1,4 +1,4 @@
-use crate::{graphic::html_canvas::HtmlCanvas, key_event::KeyEvent};
+use crate::{canvas::HtmlCanvas, key_event::KeyEvent};
 
 pub trait GameState<T>
 where
