--- src/graphic.rs
@@ -1,3 +1,0 @@
-pub mod dot;
-pub mod html_canvas;
-pub mod html_image;
