+++ src/lib.rs
@@ -1,4 +1,6 @@
+pub mod canvas;
+pub mod dot;
 pub mod game_loop;
 pub mod game_state;
-pub mod graphic;
+pub mod image;
 pub mod key_event;
